requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-cors-headers==3.5.0
psycopg2==2.8.6
requests==2.24.0
pandas==1.1.3
